Move more code out of ActivityController.
Move code to get Activity title into ShadowActivity.
Remove unused ApplicationAdapter.
